// types/interfaces.ts
type ItemResponse = {
  banner?: string;
  categoryID?: {
    _id?: string;
    categoryName?: string;
    file?: string;
  };
  description?: string;
  files?: string[];
  price?: number;
  title?: string;
  _id?: string;
  __v?: number;
  createdAt?: string;
  updatedAt?: string;
  quantity?: number;
  cost?: number;
} & { itemID?: string };
// ...
interface CartState {
  cart: {
    numberOfItems: number;
  };
}